pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "scholarly-similarity"
version = "0.1.0"
description = "Local DOI-to-similarity research harness scaffold"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Scholarly Similarity Prototype" }]
dependencies = [
  "pyyaml>=6,<7",
  "pydantic>=2,<3",
  "typer>=0.12,<1",
  "httpx>=0.27,<1",
  "streamlit>=1.43,<2",
  "plotly>=5.24,<7",
]

[project.optional-dependencies]
dev = ["pytest>=8,<10"]

[project.scripts]
scholarly-similarity = "src.cli.main:run"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
